+++ src/realpaver/Linearizer.cpp
@@ -384,7 +384,7 @@ bool LinearizerAffine::make(LPModel &lpm, const IntervalBox &B)
    }
 
    // inserts the affine forms as constraints in the linear model
-   for (int i = 0; i < creator.nbFuns(); ++i)
+   for (size_t i = 0; i < creator.nbFuns(); ++i)
    {
       const AffineForm &f = creator.fun(i);
